Refactors into a Page Class
